# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Snap3 Turbo API
  description: |
    Snap3 Turbo briefs → Veo3 preview → QA Gate → exports E2E API
    
    ## Evidence Pack v2 + Audit System
    
    All export endpoints now include comprehensive audit trails and data redaction:
    
    ### Audit Headers
    - `X-Export-SHA256`: SHA256 hash of exported data for integrity verification
    - `X-Export-Size`: Size of payload in bytes
    - `X-Export-Timestamp`: Export timestamp (ISO 8601)
    - `ETag`: Weak ETag for caching based on content hash
    
    ### Data Redaction
    VDP_FULL sensitive data is automatically redacted before export:
    - Removes: overall_analysis, asr_transcript, ocr_text, product_mentions
    - Masks: view_count, like_count, comment_count  
    - Strips: internal/*, debug/*, temp/*, raw/* paths
    
    ### Caching & ETags
    All exports support conditional requests using ETags for bandwidth optimization.
    
    ### Security
    - All exports are logged with audit trail
    - IP addresses are masked for privacy
    - User agents are sanitized 
    - No VDP_FULL data ever exposed in public APIs
  version: 2.0.0
  contact:
    name: Snap3 Team
    email: support@snap3.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.snap3.com
    description: Production server
paths:
...
  /export/brief/{id}:
    get:
      summary: Export brief
      description: Export brief with VDP_MIN + Evidence (VDP_FULL excluded) with audit trail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z0-9]{8}$'
          description: Digest ID (8 alphanumeric characters)
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, stream]
            default: json
          description: Response format (json or stream)
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
          description: ETag for conditional requests
      responses:
        '200':
          description: Brief exported with audit information
          headers:
            ETag:
              description: Evidence digest ETag for caching
              schema:
                type: string
                pattern: '^W?"[a-f0-9]{16}"$'
            X-Export-SHA256:
              description: SHA256 hash of the exported data
              schema:
                type: string
                pattern: '^[a-f0-9]{64}$'
            X-Export-Size:
              description: Size of exported data in bytes
              schema:
                type: integer
                minimum: 1
            X-Export-Timestamp:
              description: Export timestamp (ISO 8601)
              schema:
                type: string
                format: date-time
            Cache-Control:
              description: Cache control directive
              schema:
                type: string
                enum: ['private, max-age=3600', 'no-cache']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefExport'
        '304':
          description: Not Modified (ETag match)
          headers:
            ETag:
              description: Evidence digest ETag
              schema:
                type: string
            Cache-Control:
              description: Cache control directive
              schema:
                type: string
        '404':
          description: Brief not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '405':
          description: Method not allowed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /export/json/{id}:
    get:
      summary: Export JSON
      description: Export JSON with VDP_MIN + Evidence (VDP_FULL excluded) with audit trail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z0-9]{8}$'
          description: Digest ID (8 alphanumeric characters)
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
          description: ETag for conditional requests
      responses:
        '200':
          description: JSON exported with audit information
          headers:
            ETag:
              description: Evidence digest ETag for caching
              schema:
                type: string
                pattern: '^W?"[a-f0-9]{16}"$'
            X-Export-SHA256:
              description: SHA256 hash of the exported data
              schema:
                type: string
                pattern: '^[a-f0-9]{64}$'
            X-Export-Size:
              description: Size of exported data in bytes
              schema:
                type: integer
                minimum: 1
            X-Export-Timestamp:
              description: Export timestamp (ISO 8601)
              schema:
                type: string
                format: date-time
            Cache-Control:
              description: Cache control directive
              schema:
                type: string
                default: 'private, max-age=3600'
            Content-Disposition:
              description: Attachment download hint (when artifact exists)
              schema:
                type: string
                pattern: '^attachment; filename="export-[A-Z0-9]{8}.json"$'
            X-Export-Artifact-URL:
              description: Signed URL for export artifact (when exists)
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONExport'
        '304':
          description: Not Modified (ETag match)
          headers:
            ETag:
              description: Evidence digest ETag
              schema:
                type: string
            Cache-Control:
              description: Cache control directive
              schema:
                type: string
        '404':
          description: Export not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '405':
          description: Method not allowed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    VDP_MIN:
      type: object
      required:
        - id
        - title
        - description
        - category
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 500
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Evidence:
      type: object
      required:
        - id
        - type
        - content
        - confidence
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [text, url, image, video]
        content:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        metadata:
          type: object
          properties:
            source:
              type: string
            timestamp:
              type: string
              format: date-time
            tags:
              type: array
              items:
                type: string
    EvidencePack:
      type: object
      required:
        - digestId
        - trustScore
        - evidenceChips
        - synthIdDetected
      properties:
        digestId:
          type: string
        trustScore:
          type: number
          minimum: 0
          maximum: 1
          description: Trust score from 0.0 to 1.0
        evidenceChips:
          type: array
          minItems: 3
          maxItems: 5
          items:
            type: string
          description: Key evidence points (e.g., "5.2M views", "Viral", "Verified on Instagram")
        synthIdDetected:
          type: boolean
        provenance:
          type: string
          format: uri
          description: Original source URL
...
    Veo3Prompt:
      type: object
      required:
        - prompt
        - duration
        - aspect
        - resolution
      properties:
        prompt:
          type: string
          minLength: 1
          description: Text prompt for Veo3 generation
        duration:
          type: number
          const: 8
          description: Fixed duration of 8 seconds
        aspect:
          type: string
          const: '16:9'
          description: Fixed aspect ratio of 16:9
        resolution:
          type: string
          enum: ['720p', '1080p']
          description: Video resolution quality
        model:
          type: string
          default: 'veo-3'
          description: Model identifier
        parameters:
          type: object
          description: Optional model parameters
          properties:
            temperature:
              type: number
              minimum: 0
              maximum: 1
              default: 0.7
              description: Model temperature
            topP:
              type: number
              minimum: 0
              maximum: 1
              default: 0.9
              description: Top-p sampling parameter
            maxTokens:
              type: number
              minimum: 1
              maximum: 1000
              default: 500
              description: Maximum tokens
        shots:
          type: array
          description: Optional shot descriptions
          items:
            type: object
            properties:
              duration:
                type: number
              description:
                type: string
              camera:
                type: string
...
    BriefExport:
      type: object
      required:
        - digestId
        - title
        - scenes
        - evidence
        - evidencePack
      properties:
        digestId:
          type: string
          pattern: '^[A-Z0-9]{8}$'
        title:
          type: string
        scenes:
          type: array
          items:
            $ref: '#/components/schemas/TextboardScene'
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
        evidencePack:
          $ref: '#/components/schemas/EvidencePack'
    TextboardScene:
      type: object
      required:
        - role
        - durationSec
        - visual
        - audio
      properties:
        role:
          type: string
          enum: [hook, development, climax]
        durationSec:
          type: number
        visual:
          type: string
        audio:
          type: string
    JSONExport:
      type: object
      required:
        - digestId
        - data
        - exportedAt
      properties:
        digestId:
          type: string
          pattern: '^[A-Z0-9]{8}$'
        data:
          type: object
          required:
            - vdp_min
            - evidence
          properties:
            vdp_min:
              $ref: '#/components/schemas/VDP_MIN'
            evidence:
              type: array
              items:
                $ref: '#/components/schemas/Evidence'
            videoGenIR:
              $ref: '#/components/schemas/VideoGenIR'
            veo3Prompt:
              $ref: '#/components/schemas/Veo3Prompt'
            evidencePack:
              $ref: '#/components/schemas/EvidencePack'
        exportedAt:
          type: string
          format: date-time
    VideoGenIR:
      type: object
      required:
        - durationSec
        - aspect
        - resolution
        - cuts
      properties:
        durationSec:
          type: number
          const: 8
        aspect:
          type: string
          const: '16:9'
        resolution:
          type: string
          enum: ['720p', '1080p']
        cuts:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: number
              type:
                type: string
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
          description: URI reference that identifies the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          description: URI reference that identifies the specific occurrence
        code:
          type: string
          enum:
            - UNSUPPORTED_AR_FOR_PREVIEW
            - INVALID_DURATION
            - MISSING_FIRST_FRAME
            - PROVIDER_QUOTA_EXCEEDED
            - RATE_LIMITED
            - PROVIDER_POLICY_BLOCKED
            - EMBED_DENIED
            - QA_RULE_VIOLATION
            - VALIDATION_ERROR
            - BAD_REQUEST
            - RESOURCE_NOT_FOUND
            - NOT_FOUND
            - UNAUTHORIZED
            - FORBIDDEN
            - METHOD_NOT_ALLOWED
            - INTERNAL_SERVER_ERROR
            - CONFLICT
            - TIMEOUT
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
          description: Application-specific error code
        fix:
          type: string
          description: Suggested fix for the problem
        retryAfter:
          type: integer
          minimum: 1
          description: Seconds to wait before retrying
        violations:
          type: array
          description: Validation violations
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Field path that failed validation
              message:
                type: string
                description: Description of the validation failure
              code:
                type: string
                description: Validation error code
        timestamp:
          type: string
          format: date-time
          description: When the problem occurred
        traceId:
          type: string
          description: Unique identifier for request tracing
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []